+++ tests/test_ux_features.py
@@ -337,7 +337,7 @@ class TestCombinedFeatures:
 class TestVersionUpdated:
     """Test that version is correct"""
 
-    def test_version_is_3_0_11(self):
-        """Test that version is 3.0.11"""
+    def test_version_is_3_0_12(self):
+        """Test that version is 3.0.12"""
         from cja_sdr_generator import __version__
-        assert __version__ == "3.0.11"
+        assert __version__ == "3.0.12"